Move App into /lib
