+++ example/lib/App/index.js
