+++ example/src/client.js
@@ -1,6 +1,7 @@
 import React from "react";
 import DOM from "react-dom";
 
-import App from "./components/App";
+// Found in /lib via webpack's resolve.root
+import App from "App";
 
 DOM.render(<App />, document.getElementById("app"));
